WIP Get multi-page layouts and route navigation working.
