+++ src/app/sequence/page.tsx
@@ -1,0 +1,9 @@
+import React, { ReactNode } from 'react';
+
+
+export default async function Home() {
+  return (
+    <main>
+    </main>
+  );
+}
